Bulletin Board
The first photo is the beginning stages of creating the board. After talking with Ron and Alex I was able to find a general skill that I could put into steps to follow in order to create a skill like we had to for our Alexa in the State Farm lab. The second photo is our finished project with all of the information we found on the web about creating the skill.